config param for guess support, docs and minor tweaks
