+++ internal/config/config_guesser.go
@@ -1,0 +1,5 @@
+//go:build amd64 && !noguesser
+
+package config
+
+const GuessingSupported = true
