+++ internal/config/config_noguesser.go
@@ -1,0 +1,6 @@
+//go:build arm64 || noguesser
+
+package config
+
+// https://github.com/robherley/snips.sh/issues/39
+const GuessingSupported = false
